+++ src/backend/resolvers/mutations/organization/mbm/lbm/relational-resolvers/index.ts
@@ -1,6 +1,6 @@
 import serviceAndServiceCategoryResolvers from './serviceAndServiceCategory'
 import lbmAndServiceResolvers from './lbmAndService'
-// import lbmAndPersonResolvers from './lbmAndPerson'
+import lbmAndPersonResolvers from './lbmAndPerson'
 import lbmAndLbmTypeResolvers from './lbmAndLbmType'
 import lbmAndKeyEventResolvers from './lbmAndKeyEvent'
 import lbmAndPayer from './lbmAndPayer'
@@ -8,7 +8,7 @@ import lbmAndPayer from './lbmAndPayer'
 export default {
   ...serviceAndServiceCategoryResolvers,
   ...lbmAndServiceResolvers,
-  // ...lbmAndPersonResolvers,
+  ...lbmAndPersonResolvers,
   ...lbmAndLbmTypeResolvers,
   ...lbmAndKeyEventResolvers,
   ...lbmAndPayer,