+++ src/backend/resolvers/mutations/organization/mbm/lbm/relational-resolvers/lbmAndPerson/index.ts
@@ -1,0 +1,5 @@
+import connectLbmAndPerson from './replace'
+
+export default {
+  connectLbmAndPerson,
+}