+++ src/backend/resolvers/queries/organizations/mbm/lbm/joins/index.ts
@@ -1,11 +1,11 @@
-// import JOIN_lbms_people from './JOIN_lbms_people'
+import JOIN_lbms_people from './JOIN_lbms_people'
 import JOIN_lbms_payers from './JOIN_lbms_payers'
 import JOIN_lbms_lbmsServices from './JOIN_lbms_lbmsServices'
 import JOIN_lbms_lbmsTypes from './JOIN_lbms_lbmsTypes'
 import JOIN_lbmsServices_lbmsServicesCategories from './JOIN_lbmsServices_lbmsServicesCategories'
 
 export default {
-  // JOIN_lbms_people,
+  JOIN_lbms_people,
   JOIN_lbms_payers,
   JOIN_lbms_lbmsServices,
   JOIN_lbms_lbmsTypes,