+++ src/backend/typeDefs/mutations/organization/mbm/lbm/relational-typeDefs/lbmAndPerson/index.ts
@@ -1,0 +1,3 @@
+import connectLbmAndPerson from './upsert'
+
+export default [connectLbmAndPerson]