+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Build out LBM Payer Partnerships template table view [#893](https://github.com/pulse-analytics/wave-db-mgmt/pull/893)
   - Build out LBM Service Category CRUD [#894](https://github.com/pulse-analytics/wave-db-mgmt/pull/894)
   - Build out LBM Service CRUD [#895](https://github.com/pulse-analytics/wave-db-mgmt/pull/895)
+  - Build out LBM <=> People connections widget and CRUD [#899](https://github.com/pulse-analytics/wave-db-mgmt/pull/899)
+  - Build out LBM Influencers template table view [#900](https://github.com/pulse-analytics/wave-db-mgmt/pull/900)
   - Build out LBM Service <=> LBM Service Category CRUD [#896](https://github.com/pulse-analytics/wave-db-mgmt/pull/896)
   - Build out LBM <=> LBM Service CRUD [#897](https://github.com/pulse-analytics/wave-db-mgmt/pull/897)
   - Build out LBM Services VIEW page [#898](https://github.com/pulse-analytics/wave-db-mgmt/pull/898)